requirements.txt
ansible==2.1.0.0
cffi==1.7.0
cryptography==1.4
enum34==1.1.6
et-xmlfile==1.0.1
idna==2.1
ipaddress==1.0.16
jdcal==1.2
Jinja2==2.2
MarkupSafe==0.23
mechanize==0.2.5
openpyxl==2.3.5
paramiko==1.10
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
PyYAML==3.11
requests==2.10.0
six==1.10.0
slackclient==1.0.0
speedtest-cli==0.3.4
websocket-client==0.37.0
xlrd==1.0.0
